exifdata/requirements.txt
ExifRead==3.0.0
Pillow==9.2.0
